requirements.txt
python-a2s==1.3.0
bottle==0.12.25
requests==2.31.0
gunicorn==21.2.0
bottle-cors-plugin==0.1.9
